[Enhance] refactor PostCardWideHeader
