--- src/components/post-list/card/wide/header/props.d.ts
@@ -1,8 +1,0 @@
-type PostCardWideHeaderProps = {
-  title: string;
-  time: number;
-  name: string;
-  profileImageUrl: string;
-};
-
-export default PostCardWideHeaderProps;
